requirements.txt
numpy==1.26.4
scikit-learn==1.7.1
requests==2.32.5
google-genai==1.30.0
pyyaml==6.0.2
python-dotenv==1.1.1
torch==2.2.2
sentence-transformers==2.2.2
transformers==4.29.2
